WIP: board: kent: gfxboard changes (2/?)
This is patch two of ???? creating the kent_gfxboard Linux port.

This patch shows the modifications that were made to the Atmel
sama5d27_som1_ek port to represent the hardware present on the Kent
board. These files were copied from the corresponding sama5d27_som1_ek
files in the previous patch.

The functional changes are:

- TODO